requirements.txt
requests
web3
websocket-client
